+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - improvement: mixed type assertion simplified - one can assert only for string value. Exceptions are assertProperty/Attribute where additionally one can assert for boolean
 - improvement: Jest/Puppeteer project (export) has option to run in Firefox ([custom build](https://www.npmjs.com/package/puppeteer-firefox))
 - improvement: test steps and refs now support comments
+- improvement: add suite autosave option
 
 ### Updates